# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  username: string;
  role: UserRole;
  isActive: boolean;
  createdAt: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 8 */
  password: string;
  role?: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserUpdate {
  role?: UserUpdateRole;
  isActive?: boolean;
}

export interface PasswordResetInput {
  /** @minLength 8 */
  newPassword: string;
}

export interface AdminStats {
  totalUsers: number;
  totalPhotos: number;
  totalVideos: number;
  totalStorageBytes: number;
}

export type MediaUploadUrlInputMediaType = typeof MediaUploadUrlInputMediaType[keyof typeof MediaUploadUrlInputMediaType];


export const MediaUploadUrlInputMediaType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface MediaUploadUrlInput {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
  mediaType: MediaUploadUrlInputMediaType;
}

export interface MediaUploadUrlResponse {
  signature: string;
  timestamp: number;
  cloudName: string;
  apiKey: string;
  folder: string;
}

export type MediaInputMediaType = typeof MediaInputMediaType[keyof typeof MediaInputMediaType];


export const MediaInputMediaType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface MediaInput {
  /** @minLength 1 */
  originalFilename: string;
  /** @minLength 1 */
  mimeType: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  objectPath: string;
  mediaType: MediaInputMediaType;
  /** Optional folder grouping name */
  folderName?: string | null;
}

export interface MediaUpdate {
  favorite: boolean;
}

export type MediaMediaType = typeof MediaMediaType[keyof typeof MediaMediaType];


export const MediaMediaType = {
  photo: 'photo',
  video: 'video',
} as const;

export interface Media {
  id: number;
  originalFilename: string;
  mimeType: string;
  extension: string;
  size: number;
  mediaType: MediaMediaType;
  uploadedByUsername: string;
  uploadedAt: string;
  favorite: boolean;
  hasThumbnail: boolean;
  folderName?: string | null;
}

export interface MediaListResponse {
  items: Media[];
  total: number;
  hasMore: boolean;
}

export interface MediaStats {
  totalStorageBytes: number;
  storageLimitBytes: number;
  totalPhotos: number;
  totalVideos: number;
}

export type ListMediaParams = {
type?: ListMediaType;
search?: string;
favorite?: boolean;
sortBy?: ListMediaSortBy;
sortDir?: ListMediaSortDir;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};

export type ListMediaType = typeof ListMediaType[keyof typeof ListMediaType];


export const ListMediaType = {
  photo: 'photo',
  video: 'video',
} as const;

export type ListMediaSortBy = typeof ListMediaSortBy[keyof typeof ListMediaSortBy];


export const ListMediaSortBy = {
  date: 'date',
  size: 'size',
  name: 'name',
} as const;

export type ListMediaSortDir = typeof ListMediaSortDir[keyof typeof ListMediaSortDir];


export const ListMediaSortDir = {
  asc: 'asc',
  desc: 'desc',
} as const;