# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalUsers: number;
  totalPhotos: number;
  totalVideos: number;
  totalStorageBytes: number;
}