# lib/api-zod/src/generated/types/media.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MediaMediaType } from './mediaMediaType';

export interface Media {
  id: number;
  originalFilename: string;
  mimeType: string;
  extension: string;
  size: number;
  mediaType: MediaMediaType;
  uploadedByUsername: string;
  uploadedAt: Date;
  favorite: boolean;
  hasThumbnail: boolean;
}